+++ JASS/changelog.md
@@ -1,4 +1,8 @@
 # Changelogs for Perlin Noise JASS
 
+    v1.1.0
+        - Added OpenSimplex2D function
+        - Added a whitespace in function, Lerp
+
     v1.0.0
         - Initial release