+++ Lua/changelog.md
@@ -1,5 +1,9 @@
 # Changelogs for Perlin Noise Lua
 
+    v1.1.0
+        - Added Noise.openSimplex2D function
+        - Added a whitespace in the local function, lerp
+
     v1.0.1
         - Fixed wrong parameters for the function perlin1D (thanks to Drake53)
         - Fixed wrong parameters for the function grad1D